purpose-x-kill: Don't calculate 'other-buffers' unless necessary
Fixes #149
The problem is that 'gathering a list of other buffers with the same purpose' conses due to an inevitable maphash later on, and  even though it's not a ton, it adds up quick because this function is called on every kill-buffer call. Even for the tons and tons of background temp buffers used by company, helm, sly, and others.

The fix: We don't need to figure out the list of other buffers unless we actually come across a window that was displaying the killed buffer. So this fix just delays calculating it until we actually need it.