Added extra functions for helm
When calling a function to open files or switch buffer, purpose
selects what function to use depending if ido-mode is enabled.
However, it is not using specific functions if helm-mode is
enabled e.g. helm-find-files instead of find-file.

The macro purpose-ido-caller now receives an extra argument for
a function to be called in case helm-mode is enabled.