Fix for Emacs Ipython Notebooks (ein)
Ein uses uses a magic trick to give the user the impression of
navigating one single buffer, where in reality they are magically
switching between different buffers whenever they enter a markdown
cell (markwodn-mode), a Python cell (python-mode), in a different
language or in between two cells (fundamental mode).

The problematic one is the fundamental mode one, going to `general`
purpose where the rest of the cells are `edit`. This causes a glitch,
where the notebook will be opened in two windows, with the focus
switching from the one to the other as the cursor goes through the
notebook.